Rename 'annotations' to 'decorators'
